Move specs into separate directory
