+++ karma.conf.js
@@ -3,9 +3,10 @@ process.env.CHROME_BIN = require("puppeteer").executablePath();
 module.exports = function(config) {
   config.set({
     frameworks: ["jasmine", "karma-typescript"],
-    files: ["src/**/*.ts"],
+    files: ["src/**/*.ts", "spec/**/*.ts"],
     preprocessors: {
-      "**/*.ts": "karma-typescript"
+      "src/**/*.ts": ["karma-typescript", "coverage"],
+      "spec/**/*.ts": "karma-typescript"
     },
     reporters: ["progress", "karma-typescript", "notify"],
     browsers: ["ChromeHeadless"]
